Update cell_assembly_detection.py
alpha parameter in the " # testing for higher order assemblies" part gives the wrong value as it should take the original alpha=0.05 in the right side.
alpha = alpha / float(len(w2_to_test) * n_as * (2 * max_lag + 1))
As alpha in previous lines recalled and its value is changed because of same alpha values in both side of equation, it should be changed in a way that the value of alpha in the right side always is be the same is the alpha input parameter.
alpha parameter name in input parameter changed to alph in order to be constant and alpha parameter in the right side of alpha = alph/ float(len(w2_to_test) * n_as * (2 * max_lag + 1)) changes to alph. In this case, the alpha value in the right side is always constant and get the entered value in input. This simple change, will cause to get correct outputs based on the proposed algorithm.